+++ intg-sonysdcp/driver.py
@@ -93,7 +93,6 @@ async def on_r2_connect() -> None:
 
 
 @api.listens_to(ucapi.Events.DISCONNECT)
-#TODO Find out how to prevent the remote from constantly reconnecting when the integration is not running without deleting the integration configuration on the remote every time
 async def on_r2_disconnect() -> None:
     """
     Disconnect notification from the remote Two.
@@ -185,9 +184,10 @@ async def main():
 
     if config.setup.get("setup_complete"):
         if config.POLLER_INTERVAL == 0:
-            _LOG.info("POLLER_INTERVAL is " + str(config.POLLER_INTERVAL) + ". Skip creation of attributes puller task")
+            _LOG.info("POLLER_INTERVAL set to " + str(config.POLLER_INTERVAL) + ". Skip creation of attributes puller task")
         else:
             loop.create_task(attributes_poller(config.setup.get("id"), config.POLLER_INTERVAL))
+            _LOG.debug("Created attributes puller task with an interval of " + str(config.POLLER_INTERVAL) + " seconds")