+++ CHANGELOG.md
@@ -12,13 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added attributes puller function that checks all available integration attributes of the projector every 20 seconds by default. The interval can be changed in config.py. Set to 0 to deactivate this function.
 - Added runtime storage to reduce config file access load
+- Added missing SDCP error response messages in protocol.py
 
 
 ## [0.3-beta] - 2024-04-17
 
 ### Added
 - Show pySDCP exception messages in log
-- Show SDCP error respond message according to Sony's documentation instead of just the error code
+- Show SDCP error response message according to Sony's documentation instead of just the error code
 
 ### Changed
 - Split up setup flow and media player into separate files